reduce terraform.js
