+++ src/utils.js
@@ -1,8 +1,10 @@
 const fs = require('fs');
 const PATH = require('path');
+const { Buffer } = require('buffer');
 const fetch = require('node-fetch');
 const NodeSSH = require('node-ssh').NodeSSH;
 const stripAnsi = require('strip-ansi');
+const winston = require('winston');
 
 const { FileMagic, MagicFlags } = require('@npcz/magic');
 const tempy = require('tempy');
@@ -179,6 +181,41 @@ const gpuPresent = async () => {
   return gpu;
 };
 
+const tfCapture = async (command, args = [], options = {}) => {
+  return new Promise((resolve, reject) => {
+    const stderrCollection = [];
+    const tfProc = require('child_process').spawn(command, args, options);
+    tfProc.stdout.on('data', (buf) => {
+      const parse = (line) => {
+        if (line === '') return;
+        try {
+          const log = JSON.parse(line);
+          if (log['@level'] === 'error') {
+            winston.error(`terraform error: ${log['@message']}`);
+          } else {
+            winston.info(log['@message']);
+          }
+        } catch (err) {
+          // Failed to parse json from buffer
+          winston.info(line);
+        }
+      };
+      buf.toString('utf8').split('\n').forEach(parse);
+    });
+    tfProc.stderr.on('data', (buf) => {
+      stderrCollection.push(buf);
+    });
+    tfProc.on('close', (code, signal) => {
+      if (code !== 0) {
+        const stderrOutput = Buffer.concat(stderrCollection).toString('utf8');
+        reject(stderrOutput);
+      }
+      resolve();
+    });
+  });
+};
+
+exports.tfCapture = tfCapture;
 exports.exec = exec;
 exports.fetchUploadData = fetchUploadData;
 exports.upload = upload;
